+++ packages/site/cms-site/src/blocks/YouTubeVideoBlock.tsx
@@ -1,12 +1,14 @@
 "use client";
 
-import { PreviewSkeleton, type PropsWithData, useIsElementInViewport, withPreview } from "@comet/site-react";
-import clsx from "clsx";
-import { type ReactElement, type ReactNode, useRef, useState } from "react";
+import { ReactElement, ReactNode, useRef, useState } from "react";
+import styled, { css } from "styled-components";
 
-import { type YouTubeVideoBlockData } from "../blocks.generated";
-import { VideoPreviewImage, type VideoPreviewImageProps } from "./helpers/VideoPreviewImage";
-import styles from "./YouTubeVideoBlock.module.scss";
+import { YouTubeVideoBlockData } from "../blocks.generated";
+import { withPreview } from "../iframebridge/withPreview";
+import { PreviewSkeleton } from "../previewskeleton/PreviewSkeleton";
+import { useIsElementInViewport } from "./helpers/useIsElementVisible";
+import { VideoPreviewImage, VideoPreviewImageProps } from "./helpers/VideoPreviewImage";
+import { PropsWithData } from "./PropsWithData";
 
 const EXPECTED_YT_ID_LENGTH = 11;
 
@@ -106,16 +108,36 @@ export const YouTubeVideoBlock = withPreview(
                         />
                     )
                 ) : (
-                    <div
-                        ref={inViewRef}
-                        className={clsx(styles.videoContainer, fill && styles.fill)}
-                        style={!fill ? { "--aspect-ratio": aspectRatio.replace("x", "/") } : undefined}
-                    >
-                        <iframe ref={iframeRef} className={styles.youtubeContainer} allow="autoplay" src={youtubeUrl.toString()} />
-                    </div>
+                    <VideoContainer ref={inViewRef} $aspectRatio={aspectRatio.replace("x", "/")} $fill={fill}>
+                        <YouTubeContainer ref={iframeRef} allow="autoplay" src={youtubeUrl.toString()} />
+                    </VideoContainer>
                 )}
             </>
         );
     },
     { label: "Video" },
 );
+
+const VideoContainer = styled.div<{ $aspectRatio: string; $fill?: boolean }>`
+    overflow: hidden;
+    position: relative;
+
+    ${({ $aspectRatio, $fill }) =>
+        $fill
+            ? css`
+                  width: 100%;
+                  height: 100%;
+              `
+            : css`
+                  aspect-ratio: ${$aspectRatio};
+              `}
+`;
+
+const YouTubeContainer = styled.iframe`
+    position: absolute;
+    border: 0;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+`;