+++ packages/api/cms-api/CHANGELOG.md
@@ -1,5 +1,6 @@
 # @comet/cms-api
 
+<<<<<<< HEAD
 ## 8.0.0-beta.6
 
 ### Major Changes
@@ -207,6 +208,13 @@
 - 58a99bb: Fix input validation for missing child blocks
 - 7e7a4aa: Fix `title` field not added to types in `createLinkBlock`
 - f20ec6c: Make class-validator a peer dependency
+=======
+## 7.25.1
+
+### Patch Changes
+
+-   @comet/blocks-api@7.25.1
+>>>>>>> main
 
 ## 7.25.0